Unique task IDs using UUID.
Earlier, if the workload contained multiple task JSON objects with
the same name, then the task IDs for instance N would be the same
for both the tasks. The reason for this was that task IDs were
created as 'Felk-<task_name>-<task_instance>'.
Fixed this issue by appending UUID. So, the new task IDs are
'Felk-<task_name>-<task_instance>-<uuid>'